LDwebapp/backend/requirements.txt
flask
tensorflow
opencv-python-headless
pillow
numpy
